Hide the tap-to-unmute chip and unmute via the player API
Setting video.muted directly left YouTube's player state muted, so the
mobile site kept showing its "tap to unmute" chip. Clear the muted state
through the movie_player API and hide the chip in the injected CSS.
Moves injectionScript(css:) into its own file to keep the type small.
