+++ App/YouTube Player/YouTubePlayerStyles.swift
@@ -154,6 +154,8 @@ nonisolated enum YouTubePlayerStyles {
     .ytp-subtitles-button, .ytp-captions-button,
     .ytp-fullscreen-button, .ytp-size-button,
     .ytp-miniplayer-button, .ytp-pip-button,
+    .ytp-unmute, .ytp-unmute-box, .ytp-unmute-icon, .ytp-unmute-text,
+    [class*="unmute" i],
     .ytp-large-play-button, .ytp-cued-thumbnail-overlay,
     .ytp-pause-overlay, .ytp-pause-overlay-container,
     .ytp-endscreen-content, .ytp-endscreen-element,
@@ -221,45 +223,4 @@ nonisolated enum YouTubePlayerStyles {
         z-index: -1 !important;
     }
     """
-
-    static func injectionScript(css: String) -> String {
-        // Escape characters that would break out of the JS template literal.
-        let safeCSS = css
-            .replacingOccurrences(of: "\\", with: "\\\\")
-            .replacingOccurrences(of: "`", with: "\\`")
-            .replacingOccurrences(of: "${", with: "\\${")
-        return """
-        (function() {
-            function log(msg) {
-                try {
-                    window.webkit.messageHandlers.ytDebug.postMessage(String(msg));
-                } catch (e) {}
-            }
-            try {
-                var cssText = `\(safeCSS)`;
-                log('inject script start, head=' + !!document.head + ' doc=' + !!document.documentElement);
-                function inject() {
-                    if (document.getElementById('app-yt-style')) return;
-                    var parent = document.head || document.documentElement;
-                    if (!parent) { log('inject: no parent'); return; }
-                    var s = document.createElement('style');
-                    s.id = 'app-yt-style';
-                    s.textContent = cssText;
-                    parent.appendChild(s);
-                    log('inject: appended style len=' + cssText.length + ' parent=' + parent.tagName);
-                }
-                inject();
-                var observer = new MutationObserver(inject);
-                if (document.documentElement) {
-                    observer.observe(document.documentElement, { childList: true, subtree: true });
-                    log('observer attached');
-                } else {
-                    log('no documentElement for observer');
-                }
-            } catch (e) {
-                log('inject error: ' + e.message + ' stack=' + (e.stack || ''));
-            }
-        })();
-        """
-    }
 }
